# Conflicts:
#	python/paddle/pir/math_op_patch.py

+++ paddle/fluid/pir/transforms/onednn/scale_matmul_fuse_pass.cc
@@ -69,7 +69,7 @@ class ScaleMatmulFusePattern : public paddle::drr::DrrPatternBase {
     }
 
     pat.AddConstraint([&](const paddle::drr::MatchContext &match_ctx) {
-      auto scale = match_ctx.Attr<float>("scale_");
+      auto scale = match_ctx.Attr<double>("scale_");
       auto bias = match_ctx.Attr<float>("bias");
       // conditions align with fluid pass
       if (bias != 0.0f) return false;
@@ -101,7 +101,7 @@ class ScaleMatmulFusePattern : public paddle::drr::DrrPatternBase {
 
     const auto &matmul_alpha_attr = res.ComputeAttr(
         [](const paddle::drr::MatchContext &match_ctx) -> float {
-          auto scale = match_ctx.Attr<float>("scale_");
+          auto scale = match_ctx.Attr<double>("scale_");
           return scale;
         });
 
@@ -193,7 +193,7 @@ class ScaleFusedMatmulFusePattern : public paddle::drr::DrrPatternBase {
 
     pat.AddConstraint([&](const paddle::drr::MatchContext &match_ctx) {
       auto matmul_alpha = match_ctx.Attr<float>("matmul_alpha");
-      auto scale = match_ctx.Attr<float>("scale_");
+      auto scale = match_ctx.Attr<double>("scale_");
       auto bias = match_ctx.Attr<float>("bias");
       // conditions align with fluid pass
       if (matmul_alpha == 0.0f) return false;
@@ -226,7 +226,7 @@ class ScaleFusedMatmulFusePattern : public paddle::drr::DrrPatternBase {
 
     const auto &matmul_alpha_attr = res.ComputeAttr(
         [](const paddle::drr::MatchContext &match_ctx) -> float {
-          auto scale = match_ctx.Attr<float>("scale_");
+          auto scale = match_ctx.Attr<double>("scale_");
           auto matmul_alpha = match_ctx.Attr<float>("matmul_alpha");
           return scale * matmul_alpha;
         });
